Move DFS and BFS to search.lua
