+++ graph.lua
@@ -4,8 +4,6 @@ local Object = require "object"
 
 local Neighbourhood = require "nbhood"
 
-local Queue = require "queue"
-
 -----------------------
 -- Class
 -----------------------
@@ -128,90 +126,9 @@ function Graph:iterEdges(v)
     end
 end
 
--- Do a depth-first search on the graph from vertex `s`
--- Returns a DFS tree where each vertex has a "ref" field
--- that points to the original one
--- If vertex `s` is not in the graph, returns nil
--- Parameters
---   s   : Vertex - first vertex to be visited
--- Return values
---   [1] : Graph - DFS tree
---   [2] : Vertex - vertex in DFS tree that references `s`
-function Graph:dfs(s)
-    if self:hasVertex(s) then
-        local tree = Graph:new()
-        local visited = {}
-        local ts = assert(self:_dfsVisit(tree, visited, s))
-        return tree, ts
-    end
-end
-
--- Do a breadth-first search on the graph from vertex `s`
--- Returns a BFS tree where each vertex has a "ref" field
--- that points to the original one
--- If vertex `v` is not in the graph, returns nil
--- Parameters
---   s   : Vertex - first vertex to be visited
--- Return values
---   [1] : Graph - BFS tree
---   [2] : Vertex - vertex in DFS tree that references `v`
-function Graph:bfs(s)
-    if self:hasVertex(s) then
-        local tree = Graph:new()
-        local treeVertices = {} -- (GraphVertex -> TreeVertex)
-        local treeVertexQueue = Queue:new() -- for TreeVertex only
-        local ts = tree:_addRef(s)
-        treeVertices[s] = ts
-        assert(treeVertexQueue:enqueue(ts))
-        while not treeVertexQueue:isEmpty() do
-            local tv = assert(treeVertexQueue:dequeue())
-            local v = tv.ref
-            for w, e in self:iterEdges(v) do
-                local tw = treeVertices[w]
-                if tw == nil then
-                    tw = tree:_addRef(w)
-                    treeVertices[w] = tw
-                    assert(tree:addEdge(tv, tw))
-                    assert(treeVertexQueue:enqueue(tw))
-                end
-            end
-        end
-        return tree, ts
-    end
-end
-
 -----------------------
 -- Private functions
 -----------------------
-
--- Visit vertex `v` in a depth-first search
--- If `v` has been visited already, returns nil
--- Parameters
---   tree : Graph - DFS tree
---   visited : (Vertex -> bool)
---   v : Vertex
--- Return values
---   [1] : Vertex - vertex in tree that references `v`
-function Graph:_dfsVisit(tree, visited, v)
-    if not visited[v] then
-        local tv = tree:_addRef(v)
-        visited[v] = true
-        for w, e in self:iterEdges(v) do
-            local tw = self:_dfsVisit(tree, visited, w)
-            if tw ~= nil then
-                assert(tree:addEdge(tv, tw))
-            end
-        end
-        return tv
-    end
-end
-
--- Adds a vertex with "ref" pointing to `v`
-function Graph:_addRef(v)
-    local vref = self:addVertex()
-    vref.ref = v
-    return vref
-end
 
 -----------------------
 -- Return class
