+++ test/benches.lua
@@ -4,5 +4,6 @@ return {
     "test.graph",
     "test.object",
     "test.queue",
+    "test.search",
     "test.select",
 }
